requirements.txt
aiohttp==3.6.2
alembic==1.4.2
aniso8601==7.0.0
appdirs==1.4.3
APScheduler==3.6.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
black==19.10b0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
clic==0.1.3
Click==7.0
colorama==0.4.3
coverage==5.0.4
decorator==4.4.2
docopt==0.6.2
elasticsearch==7.6.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
flake8-black==0.1.1
flake8-isort==2.9.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-GraphQL==2.0.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
fuzzywuzzy==0.18.0
gevent==20.4.0
google==2.0.3
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==0.4.15
grpcio==1.28.1
gunicorn==20.0.4
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
multidict==4.7.5
mysqlclient==1.4.6
numpy==1.18.2
packaging==20.3
pandas==1.0.3
parso==0.7.0
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pip-api==0.0.14
pipreqs==0.4.10
pluggy==0.13.1
plumbum==1.6.9
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.11.3
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.8.1
py-spy==0.3.3
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
pyhumps==1.3.1
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.16.0
pytest==5.4.1
pytest-cov==2.8.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-Levenshtein==0.12.0
pytz==2019.3
PyYAML==5.3.1
ray==0.8.4
redis==3.4.1
regex==2020.2.20
requests==2.23.0
rpyc==4.1.4
Rx==1.6.1
scipy==1.4.1
singledispatch==3.4.0.3
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
testfixtures==6.14.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.0
wrapt==1.11.2
yarg==0.1.9
yarl==1.4.2
